added new "skip_ip" option
when using bridged networking (using the "interface" option) it
is difficult for tart to obtain the IP, it will attempt it via
the hosts ARP table however this only works if the VM is making
ARP requests. So with the right software/settings it should work
however if you dont need to have the IP in terraform it might
be easier to just skip returning the IP altogether.

Then you should find the IP in your DHCP server, or you could
have the VM broadcast it, join a VPN or Tailscale network, etc